Move restriction classes to dedicated module
