+++ tests/test_token.py
@@ -1,377 +1,18 @@
 from __future__ import annotations
 
-import dataclasses
-import datetime
 import inspect
 
 import pymacaroons
 import pytest
 
-from pypitoken import exceptions, token
-
-
-def test__Restriction__dump_json():
-    class MyRestriction(token.Restriction):
-        def dump(self):
-            return {"a": ["b"]}
-
-    assert MyRestriction().dump_json() == '{"a": ["b"]}'
-
-
-def test__Restriction__load_value__pass():
-    @dataclasses.dataclass
-    class MyRestriction(token.Restriction):
-        version: int
-
-        @staticmethod
-        def _get_schema():
-            return {
-                "type": "object",
-                "properties": {
-                    "version": {"type": "integer", "const": 42},
-                },
-                "required": ["version"],
-            }
-
-        @classmethod
-        def _extract_kwargs(cls, value):
-            return {"version": value["version"]}
-
-    assert MyRestriction._load_value(value={"version": 42}).version == 42
-
-
-@pytest.mark.parametrize(
-    "value",
-    [
-        [],
-        {},
-        {"some": "key"},
-        {"version": "42"},
-        {"version": 17},
-    ],
-)
-def test__Restriction__load_value__fail(value):
-    class MyRestriction(token.Restriction):
-        @staticmethod
-        def _get_schema():
-            return {
-                "type": "object",
-                "properties": {
-                    "version": {"type": "integer", "const": 42},
-                },
-                "required": ["version"],
-            }
-
-    with pytest.raises(exceptions.LoaderError):
-        MyRestriction._load_value(value=value)
-
-
-def test__LegacyNoopRestriction__load_value__pass():
-    tok = token.LegacyNoopRestriction._load_value(
-        value={"version": 1, "permissions": "user"}
-    )
-    assert tok == token.LegacyNoopRestriction()
-
-
-@pytest.mark.parametrize(
-    "value",
-    [
-        {"permissions": "user"},
-        {"version": 2, "permissions": "user"},
-        {"version": 1},
-        {"version": 1, "permissions": {"projects": ["a"]}},
-        {"version": 1, "permissions": "something else"},
-        {"version": 1, "permissions": "user", "additional": "key"},
-    ],
-)
-def test__LegacyNoopRestriction__load_value__fail(value):
-    with pytest.raises(exceptions.LoaderError):
-        token.LegacyNoopRestriction._load_value(value=value)
-
-
-def test__LegacyNoopRestriction__extract_kwargs():
-    noop = token.LegacyNoopRestriction._extract_kwargs(value={"any": "content"})
-    assert noop == {}
-
-
-def test__LegacyNoopRestriction__check():
-    noop = token.LegacyNoopRestriction()
-    assert noop.check(context=token.Context(project="foo")) is None
-
-
-def test__LegacyNoopRestriction__dump():
-    noop = token.LegacyNoopRestriction()
-    assert noop.dump() == {"version": 1, "permissions": "user"}
-
-
-def test__LegacyNoopRestriction__from_parameters__empty():
-    assert (
-        token.LegacyNoopRestriction.from_parameters() == token.LegacyNoopRestriction()
-    )
-
-
-def test__LegacyNoopRestriction__from_parameters__not_empty():
-    assert token.LegacyNoopRestriction.from_parameters(a=1) is None
-
-
-@pytest.mark.parametrize(
-    "value, restriction",
-    [
-        (
-            {"version": 1, "permissions": {"projects": []}},
-            token.LegacyProjectsRestriction(projects=[]),
-        ),
-        (
-            {"version": 1, "permissions": {"projects": ["a"]}},
-            token.LegacyProjectsRestriction(projects=["a"]),
-        ),
-        (
-            {"version": 1, "permissions": {"projects": ["a", "b"]}},
-            token.LegacyProjectsRestriction(projects=["a", "b"]),
-        ),
-    ],
-)
-def test__LegacyProjectsRestriction__load_value__pass(value, restriction):
-    assert token.LegacyProjectsRestriction._load_value(value=value) == restriction
-
-
-@pytest.mark.parametrize(
-    "value",
-    [
-        {"permissions": {"projects": ["a"]}},
-        {"version": 2, "permissions": {"projects": ["a"]}},
-        {"version": 1},
-        {"version": 1, "permissions": "user"},
-        {"version": 1, "permissions": {"a": "b"}},
-        {"version": 1, "permissions": {"projects": "a"}},
-        {"version": 1, "permissions": {"projects": [1]}},
-        {"version": 1, "permissions": {"projects": ["a", 1]}},
-        {"version": 1, "permissions": {"projects": ["a"]}, "additional": "key"},
-        {"version": 1, "permissions": {"projects": ["a"], "additional": "key"}},
-    ],
-)
-def test__LegacyProjectsRestriction__load_value__fail(value):
-    with pytest.raises(exceptions.LoaderError):
-        token.LegacyProjectsRestriction._load_value(value=value)
-
-
-def test__LegacyProjectsRestriction__extract_kwargs():
-    value = {"version": 1, "permissions": {"projects": ["a", "b"]}}
-    kwargs = token.LegacyProjectsRestriction._extract_kwargs(value=value)
-    assert kwargs == {"projects": ["a", "b"]}
-
-
-def test__LegacyProjectsRestriction__check__pass():
-    restriction = token.LegacyProjectsRestriction(projects=["a", "b"])
-    assert restriction.check(context=token.Context(project="a")) is None
-
-
-def test__LegacyProjectsRestriction__check__fail():
-    restriction = token.LegacyProjectsRestriction(projects=["a", "b"])
-    with pytest.raises(exceptions.ValidationError):
-        restriction.check(context=token.Context(project="c"))
-
-
-def test__LegacyProjectsRestriction__dump():
-    restriction = token.LegacyProjectsRestriction(projects=["a", "b"])
-    assert restriction.dump() == {
-        "version": 1,
-        "permissions": {"projects": ["a", "b"]},
-    }
-
-
-def test__LegacyProjectsRestriction__from_parameters__empty():
-    assert token.LegacyProjectsRestriction.from_parameters() is None
-
-
-def test__LegacyProjectsRestriction__from_parameters__not_empty():
-    assert token.LegacyProjectsRestriction.from_parameters(
-        projects=["a", "b"]
-    ) == token.LegacyProjectsRestriction(projects=["a", "b"])
-
-
-def test__LegacyDateRestriction__load_value__pass():
-    assert token.LegacyDateRestriction._load_value(
-        value={"nbf": 1_234_567_890, "exp": 1_234_567_900}
-    ) == token.LegacyDateRestriction(not_before=1_234_567_890, not_after=1_234_567_900)
-
-
-@pytest.mark.parametrize(
-    "value",
-    [
-        {},
-        {"nbf": "2000-01-01 00:00:00", "exp": "2100-01-01 00:00:00"},
-        {"nbf": "1_234_567_890", "exp": "1_234_567_900"},
-        {"nbf": 1_234_567_890},
-        {"exp": 1_234_567_890},
-    ],
-)
-def test__LegacyDateRestriction__load_value__fail(value):
-    with pytest.raises(exceptions.LoaderError):
-        token.LegacyDateRestriction._load_value(value=value)
-
-
-def test__LegacyDateRestriction__extract_kwargs():
-    value = {"nbf": 1_234_567_890, "exp": 1_234_567_900}
-    kwargs = token.LegacyDateRestriction._extract_kwargs(value=value)
-    assert kwargs == {"not_before": 1_234_567_890, "not_after": 1_234_567_900}
-
-
-def test__LegacyDateRestriction__check__pass():
-    restriction = token.LegacyDateRestriction._load_value(
-        value={"nbf": 1_234_567_890, "exp": 1_234_567_900}
-    )
-    assert (
-        restriction.check(context=token.Context(project="a", now=1_234_567_895)) is None
-    )
-
-
-@pytest.mark.parametrize(
-    "value",
-    [
-        1_234_567_000,
-        1_234_568_000,
-    ],
-)
-def test__LegacyDateRestriction__check__fail(value):
-    restriction = token.LegacyDateRestriction._load_value(
-        value={"nbf": 1_234_567_890, "exp": 1_234_567_900}
-    )
-    with pytest.raises(exceptions.ValidationError):
-        restriction.check(context=token.Context(project="a", now=value))
-
-
-def test__LegacyDateRestriction__dump():
-    restriction = token.LegacyDateRestriction._load_value(
-        value={"nbf": 1_234_567_890, "exp": 1_234_567_900}
-    )
-    assert restriction.dump() == {
-        "nbf": 1_234_567_890,
-        "exp": 1_234_567_900,
-    }
-
-
-def test__LegacyDateRestriction__from_parameters__empty():
-    assert token.LegacyDateRestriction.from_parameters() is None
-
-
-@pytest.mark.parametrize(
-    "kwargs",
-    [
-        {"not_before": 1_234_567_890},
-        {"not_after": 1_234_567_900},
-        {
-            "not_before": datetime.datetime(2000, 1, 1),
-            "not_after": datetime.datetime(2100, 1, 1),
-        },
-    ],
-)
-def test__LegacyDateRestriction__from_parameters__fail(kwargs):
-    with pytest.raises(exceptions.InvalidRestriction):
-        assert token.LegacyDateRestriction.from_parameters(**kwargs) is None
-
-
-@pytest.mark.parametrize(
-    "kwargs, expected",
-    [
-        (
-            {"not_before": 1_234_567_890, "not_after": 1_234_567_900},
-            token.LegacyDateRestriction._load_value(
-                value={"nbf": 1_234_567_890, "exp": 1_234_567_900}
-            ),
-        ),
-        (
-            {
-                "not_before": datetime.datetime(
-                    2000, 1, 1, tzinfo=datetime.timezone.utc
-                ),
-                "not_after": datetime.datetime(
-                    2100, 1, 1, tzinfo=datetime.timezone.utc
-                ),
-            },
-            token.LegacyDateRestriction._load_value(
-                value={"nbf": 946_684_800, "exp": 4_102_444_800}
-            ),
-        ),
-    ],
-)
-def test__LegacyDateRestriction__from_parameters__ok(kwargs, expected):
-    assert token.LegacyDateRestriction.from_parameters(**kwargs) == expected
-
-
-def test__Restriction__get_subclasses():
-    # This test ensures we didn't forget to add new restriction classes to
-    # the set.
-    assert set(token.Restriction._get_subclasses()) == {
-        cls
-        for cls in token.Restriction.__subclasses__()
-        if cls.__module__ == "pypitoken.token"
-    }
-
-
-def test__Restriction__json_load_caveat__pass():
-    assert token.Restriction._json_load_caveat('{"a": "b"}') == {"a": "b"}
-
-
-def test__Restriction__json_load_caveat__fail():
-    with pytest.raises(exceptions.LoaderError) as exc_info:
-        token.Restriction._json_load_caveat(caveat='{"a": "b"')
-    assert (
-        str(exc_info.value) == "Error while loading caveat: "
-        "Expecting ',' delimiter: line 1 column 10 (char 9)"
-    )
-
-
-@pytest.mark.parametrize(
-    "caveat, output",
-    [
-        (
-            {"version": 1, "permissions": "user"},
-            token.LegacyNoopRestriction(),
-        ),
-        (
-            {"version": 1, "permissions": {"projects": ["a", "b"]}},
-            token.LegacyProjectsRestriction(projects=["a", "b"]),
-        ),
-    ],
-)
-def test__Restriction__load__pass(caveat, output):
-    assert token.Restriction.load(caveat=caveat) == output
-
-
-def test__Restriction__load__fail():
-    with pytest.raises(exceptions.LoaderError) as exc_info:
-        token.Restriction.load(caveat={"version": 1, "permissions": "something"})
-    assert (
-        str(exc_info.value)
-        == "Could not find matching Restriction for {'version': 1, 'permissions': 'something'}"
-    )
-
-
-def test__Restriction__load_json():
-    restriction = token.Restriction.load_json(
-        caveat='{"version": 1, "permissions": "user"}'
-    )
-    assert restriction == token.LegacyNoopRestriction()
-
-
-def test__Restriction__restrictions_from_parameters():
-    restrictions = list(
-        token.Restriction.restrictions_from_parameters(
-            projects=["a", "b"], not_before=1, not_after=5
-        )
-    )
-    assert restrictions == [
-        token.LegacyProjectsRestriction(projects=["a", "b"]),
-        token.LegacyDateRestriction(not_before=1, not_after=5),
-    ]
+from pypitoken import exceptions, restrictions, token
 
 
 def test__Token__check_caveat__pass():
     errors = []
     value = token.Token._check_caveat(
         caveat='{"version": 1, "permissions": {"projects": ["a", "b"]}}',
-        context=token.Context(project="a"),
+        context=restrictions.Context(project="a"),
         errors=errors,
     )
     assert value is True
@@ -381,7 +22,7 @@ def test__Token__check_caveat__pass():
 def test__check_caveat__fail_load_json():
     errors = []
     value = token.Token._check_caveat(
-        "{", context=token.Context(project="a"), errors=errors
+        "{", context=restrictions.Context(project="a"), errors=errors
     )
     assert value is False
     messages = [str(e) for e in errors]
@@ -395,7 +36,7 @@ def test__check_caveat__fail_load_json():
 def test__check_caveat__fail_load():
     errors = []
     value = token.Token._check_caveat(
-        '{"version": 13}', context=token.Context(project="a"), errors=errors
+        '{"version": 13}', context=restrictions.Context(project="a"), errors=errors
     )
     assert value is False
     messages = [str(e) for e in errors]
@@ -407,7 +48,7 @@ def test__check_caveat__fail_check():
 
     value = token.Token._check_caveat(
         '{"version": 1, "permissions": {"projects": ["a", "b"]}}',
-        context=token.Context(project="c"),
+        context=restrictions.Context(project="c"),
         errors=errors,
     )
     assert value is False
@@ -582,9 +223,9 @@ def test__Token__restrictions(create_token):
     tok.restrict(projects=["a", "b"])
     tok.restrict(projects=["a", "d"])
     assert tok.restrictions == [
-        token.LegacyNoopRestriction(),
-        token.LegacyProjectsRestriction(projects=["a", "b"]),
-        token.LegacyProjectsRestriction(projects=["a", "d"]),
+        restrictions.LegacyNoopRestriction(),
+        restrictions.LegacyProjectsRestriction(projects=["a", "b"]),
+        restrictions.LegacyProjectsRestriction(projects=["a", "d"]),
     ]
